requirements/base.txt
django==5.0.6
django-environ==0.11.2
django-autoslug==1.9.9
django-cachalot==2.6.2
django-countries==7.6.1
django-filter==24.2
django-cleanup==8.1.0
django-redis==5.4.0
django-colorfield==0.11.0
django-shortuuidfield==0.1.3
django-cors-headers==4.3.1
django-model-utils==4.5.1
django-celery-email==3.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
pillow==10.3.0
djoser==2.2.3
social-auth-app-django==5.4.1
argon2-cffi==23.1.0
pytz==2024.1
redis==5.0.5
celery==5.4.0
flower==2.0.1
gunicorn==22.0.0
coverage==7.5.3
